Moved Frame struct into element context, to be close to it's associated methods
